refactor(heroRoller): move testing helpers into new subsystem
