+++ module/heroRoller/dice-testing-helper.mjs
